#134 converting index to typescript
