+++ package.json
@@ -1,21 +1,23 @@
 {
   "name": "python-shell",
-  "version": "0.5.0",
+  "version": "1.0.0",
   "description": "Run Python scripts from Node.js with simple (but efficient) inter-process communication through stdio",
   "keywords": [
     "python"
   ],
   "scripts": {
-    "test": "mocha -R spec",
-    "appveyorTest": "mocha --ui tdd --reporter mocha-appveyor-reporter test/*.js"
+    "test": "tsc -p ./ && mocha -R spec",
+    "appveyorTest": "mocha --ui tdd --reporter mocha-appveyor-reporter test/*.js",
+    "compile":"tsc -watch -p ./"
   },
   "dependencies": {},
   "devDependencies": {
     "@types/mocha": "^2.2.48",
     "@types/node": "^9.3.0",
-    "should": "^13.2.1",
     "mocha": "^5.0.0",
-    "mocha-appveyor-reporter": "^0.4.0"
+    "mocha-appveyor-reporter": "^0.4.0",
+    "should": "^13.2.1",
+    "typescript": "^3.0.1"
   },
   "repository": {
     "type": "git",
